google-drive/pyproject.toml
[project]
name = "google-drive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.1",
    "pandas>=2.2.3",
    "pydantic>=2.11.5",
    "pytz>=2025.2",
]
